Update component-stack to use fixtures
