+++ test/development/acceptance-app/fixtures/component-stack/app/component.js
@@ -1,0 +1,9 @@
+'use client'
+const isClient = typeof window !== 'undefined'
+export default function Component() {
+  return (
+    <div>
+      <p>{isClient ? 'client' : 'server'}</p>
+    </div>
+  )
+}
