+++ test/development/acceptance-app/fixtures/component-stack/app/page.js
@@ -1,0 +1,8 @@
+import Component from './component'
+export default function Mismatch() {
+  return (
+    <main>
+      <Component />
+    </main>
+  )
+}
